requirements.txt
reflex==0.8.26
pytesseract
pillow
reflex-enterprise
google-genai